Separate FileInstance from fs:fs
Bug: b/544860009
